+++ src/sections/crosshair.js
@@ -19,6 +19,8 @@ export default class CrosshairSection extends Section {
 		this._waitUntilFinished = true;
 	}
 
+	static niceName = "Crosshair";
+
 	/**
 	 * Sets the type of MeasurableTemplate to create, see CONST.MEASURED_TEMPLATE_TYPES
 	 */
@@ -175,9 +177,11 @@ export default class CrosshairSection extends Section {
 			showRange: CrosshairsDocument.defaultConfig.location.showRange,
 			lockToEdge: CrosshairsDocument.defaultConfig.location.lockToEdge,
 			lockToEdgeDirection: CrosshairsDocument.defaultConfig.location.lockToEdgeDirection,
-			offset: CrosshairsDocument.defaultConfig.location.offset
+			offset: CrosshairsDocument.defaultConfig.location.offset,
+			wallBehavior: CrosshairsDocument.defaultConfig.location.wallBehavior
 		}, inOptions,);
 		inLocation = this._validateLocation(inLocation);
+		if (typeof inOptions.wallBehavior !== "string") throw this.sequence._customError(this, "wallBehavior", "inOptions.wallBehavior must be of type string");
 		if (inLocation === undefined) throw this.sequence._customError(this, "location", "could not find position of given object");
 		if (inOptions.limitMinRange && !lib.is_real_number(inOptions.limitMinRange)) throw this.sequence._customError(this, "lockLocation", "inOptions.limitMinRange must be of type number");
 		if (inOptions.limitMaxRange && !lib.is_real_number(inOptions.limitMaxRange)) throw this.sequence._customError(this, "lockLocation", "inOptions.limitMaxRange must be of type number");
@@ -200,7 +204,8 @@ export default class CrosshairSection extends Section {
 			lockToEdge: inOptions.lockToEdge,
 			lockToEdgeDirection: inOptions.lockToEdgeDirection,
 			showRange: inOptions.showRange,
-			offset: inOptions.offset
+			offset: inOptions.offset,
+			wallBehavior: inOptions.wallBehavior
 		};
 		return this;
 	}
@@ -219,7 +224,7 @@ export default class CrosshairSection extends Section {
 	 */
 	gridHighlight(inBool = true) {
 		if (typeof inBool !== "boolean") throw this.sequence._customError(this, "gridHighlight", "inBool must be of type boolean");
-		this.config['gridHighlight'] = inBool;
+		this._config['gridHighlight'] = inBool;
 		return this;
 	}